go.mod: Use a recent client-go version
When trying to update the runtime-specs to a new version I needed to run
"go mod vendor" that broke the compilation in horrible ways.

We were using an old client-go version that was marked as
"+incompatible". From the client-go website, it seems those versions are
known to cause the exact same issues I run into when using "go mod
vendor" and "go mod tidy":
	https://github.com/kubernetes/client-go/blob/master/INSTALL.md#troubleshooting

From the link, I was hitting the issue "Conflicting requirements for
older client-go versions".

This commit just updates client-go to use a recent version not marked as
"+incomptabile". The only code change needed was to use a context, I
trivially used "context.TODO()" for that.

To keep things in order, I also run:
	go mod tidy
	go mod vendor