Use lastest OCI runtime-spec
The runtime-spec changes were already merged in this PR:
        https://github.com/opencontainers/runtime-spec/pull/1074

To reference the new merge commit with the latest fixes I just run:
	go get 'github.com/opencontainers/runtime-spec@9c848d91e8cf872e7453296832d66de6325e1e25
        go mod vendor

And commited the changes.

I also update the code to match the new spec. While doing that, I
rewrote the code to close fds in _most_ cases. It is tricky to close
them before we have a reference to the `fds` slice, so that is left as a
follow-up improvement.

Signed-off-by: Rodrigo Campos <rodrigo@kinvolk.io>